Added job file option
Added a commandline tool option "-j job.json" to load a job file with
all of the commandline options as a JSON configuration alternative and
in support of other options. Included a configDefaults.json file as an
example. Modified modules.js to accept multiple URL specs. Updated
README.md, respectively.